refactor: SecurityConfig 위치 변경
