+++ src/main/java/com/kosa/fillinv/member/dto/SignUpDto.java
@@ -2,7 +2,6 @@ package com.kosa.fillinv.member.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
 import lombok.*;
 
 @Getter
